requirements.txt
pydantic==2.10.3
pandas==2.2.3
numpy==2.1.3
psycopg==3.2.3
requests==2.32.3
psycopg2--binary
tqdm==4.67.1
haversine==2.9